webdav: fix mkdir on bitrix24 - fixes #3513
Before this change bitrix24 would fail when trying to mkdir a
directory giving a 409 error.  This is somewhat RFC incompatible since
the RFC says 409 means the parent needs creating, so as a workaround
check to see if the directory really exists.  This won't slow down
compliant servers except in the unusual case of having to create
multiple layers of directories.